Refactor the URDF sub-package
NOT TESTED YET
