+++ urdfpy/__init__.py
@@ -2,7 +2,7 @@ import os
 if 'DISPLAY' not in os.environ or len(os.environ['DISPLAY']) == 0:
     os.environ['PYOPENGL_PLATFORM'] = 'egl'
 
-from .urdf import (URDFType,
+from urdfpy.urdf import (URDFType,
                    Box, Cylinder, Sphere, Mesh, Geometry,
                    Texture, Material,
                    Collision, Visual, Inertial,
